+++ test/js/spec/chai-jq.spec.js
@@ -233,7 +233,8 @@ define(["jquery", "chai"], function ($, chai) {
           .to.not.be.$visible;
       });
 
-      it("verifies visibility if parent is hidden", function () {
+      // JsDom doesn't work for parent visibility.
+      it.skipNode("verifies visibility if parent is hidden", function () {
         var $fixture = this.$fixture;
 
         $fixture.wrap("<div />");
@@ -419,28 +420,45 @@ define(["jquery", "chai"], function ($, chai) {
 
     describe("$css", function () {
       beforeEach(function () {
-        this.$fixture = $("<div style=\"width: 50px;" +
-          "border: 1px dotted black;\" />").appendTo(this.$base);
+        this.$fixture = $(
+          "<div style=\"width: 50px; border: 1px dotted black;\" />"
+        ).appendTo(this.$base);
       });
 
-      it("matches CSS property", function () {
+      it("matches explicit CSS properties", function () {
         var $fixture = this.$fixture;
 
         expect($fixture)
           .to.have.$css("width", "50px").and
+          .to.not
+            .have.$css("width", "100px").and
+            .have.$css("height", "50px");
+
+        expect($fixture)
+          .to.contain.$css("border", "1px dotted").and
+          .to.contain.$css("border", "dotted").and
+          .to.not
+            .contain.$css("border", "solid").and
+            .contain.$css("border", "2px");
+
+        expect($("<p style=\"float: left; display: none;\" />"))
+          .to.have.$css("float", "left").and
+          .to.have.$css("display", "none");
+      });
+
+      // JsDom doesn't work properly with computed properties.
+      it.skipNode("matches computed CSS properties", function () {
+        var $fixture = this.$fixture;
+
+        expect($fixture)
           .to.have.$css("border-top-style", "dotted").and
           .to.not
-          .have.$css("width", "100px").and
-          .have.$css("border-top-style", "solid");
+            .have.$css("border-top-style", "solid");
 
         expect(function () {
           expect($fixture).to.have.$css("width", "100px");
         }).to.throw("expected [object Object] to have css('width') '100px' " +
                     "but found '50px'");
-
-        expect($("<p style=\"float: left; display: none;\" />"))
-          .to.have.$css("float", "left").and
-          .to.have.$css("display", "none");
       });
     });